feat(message): cleanup client message

+++ src/main.rs
@@ -147,7 +147,7 @@ pub struct MessageInternalMineSuccess {
     coal_config: Option<coal_api::state::Config>,
     multiplier: f64,
     submissions: HashMap<Pubkey, InternalMessageSubmission>,
-    guild_total_stake: u64,
+    guild_total_stake: f64,
     guild_multiplier: f64,
     tool_multiplier: f64,
 }
